requirements.txt
Flask
Werkzeug
requests
ffmpeg-python
openai-whisper==20231117
gunicorn
APScheduler
srt
numpy==1.24.3
torch==2.1.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage
google-cloud-secret-manager>=2.0.0
psutil
boto3
Pillow
matplotlib
jsonschema